+++ main/services/category.service.js
@@ -1,4 +1,5 @@
 const Category = require('../models/category');
+const OrderCategory = require('../models/orderCategory');
 
 exports.createCategory = async (data) => {
   const category = new Category(data);
@@ -10,7 +11,12 @@ exports.updateCategory = async (id, data) => {
 };
 
 exports.getAllCategories = async () => {
-  return await Category.find();
+  // Return categories with `inUse` flag when referenced by any OrderCategory
+  const categories = await Category.find().lean();
+  return await Promise.all(categories.map(async (c) => {
+    const count = await OrderCategory.countDocuments({ category: c._id });
+    return { ...c, inUse: count > 0 };
+  }));
 };
 
 exports.getCategoryById = async (id) => {
@@ -18,5 +24,10 @@ exports.getCategoryById = async (id) => {
 };
 
 exports.deleteCategory = async (id) => {
+  // Prevent deletion if this category is referenced by any order suborders
+  const inUse = await OrderCategory.countDocuments({ category: id });
+  if (inUse > 0) {
+    throw new Error('This category is in use and cannot be deleted');
+  }
   return await Category.findByIdAndDelete(id);
 };
